requirements.txt
# YOLO框架（含训练、导出、推理）
ultralytics>=8.3.0

# ONNX推理（CPU加速，无需PyTorch运行时）
onnxruntime>=1.17.0

# 图像处理
opencv-python>=4.9.0
Pillow>=10.0.0

# 配置文件
PyYAML>=6.0

# 数值计算
numpy>=1.24.0

# 数据增强（可选，ultralytics已内置基础增强）
albumentations>=1.4.0

